requirements.txt
djangorestframework-3.8.2
pillow
python3
bootstrap3